+++ models/avoir.js
@@ -5,8 +5,8 @@ module.exports = (sequelize, DataTypes) => {
     {
       num_avoir: DataTypes.STRING,
       date_avoir: DataTypes.STRING,
-      montant_ht: DataTypes.INTEGER,
-      montant_ttc: DataTypes.INTEGER,
+      montant_ht: DataTypes.FLOAT,
+      montant_ttc: DataTypes.FLOAT,
       active: DataTypes.BOOLEAN
     },
     {}
