+++ models/facture.js
@@ -8,11 +8,11 @@ module.exports = (sequelize, DataTypes) => {
       num_document_transport: DataTypes.STRING,
       num_facture: DataTypes.STRING,
       date_facture: DataTypes.STRING,
-      montant_ht: DataTypes.INTEGER,
-      montant_ttc: DataTypes.INTEGER,
+      montant_ht: DataTypes.FLOAT,
+      montant_ttc: DataTypes.FLOAT,
       echeance_facture: DataTypes.STRING,
-      taux_applicable: DataTypes.INTEGER,
-      intérets_capitalises: DataTypes.INTEGER,
+      taux_applicable: DataTypes.FLOAT,
+      intérets_capitalises: DataTypes.FLOAT,
       active: DataTypes.BOOLEAN
     },
     {}
